Added option for osd failure domain
Allow the failure domain to be set to osd similar to the mon count
setting. This sets the CRUSH rule 0, default erasure profile, and makes
the related_osd count osds in place of relations. Finally, when in use
the first time status is run after passing the required osd count
clients are notified so any pending relations can continue.

Change-Id: I1b185b9f54ea9e15afe771e80fea10fb242da3c6
Closes-Bug: #1740961